Always use the logger for debugging output
The pysmi library uses a proper logging facility to generate its
debugging output in all cases, except for one location, which was
changed by commit git-0b2378ec to use print(). The result of that
change is that users of the library can no longer disable the output
from that debugging statement, thereby requiring work-arounds whenever
clean output is needed from the library.

This commit changes back the debugging statement to use the logging
facility as before.