Fix augmentations depending on future definitions
An AUGMENTS clause causes the pysnmp code generation to issue a call on
the object being augmented, from the context of the object performing
the augumentation. However, if the former object is defined later on in
the MIB, it does not yet exist when being augmented. The result is that
loading the generated code fails on a NameError exception.

This commit fixes the problem by issuing the augmentation-related calls
in a second pass, after all the objects have been defined. During the
second pass, the object being augmented is guaranteed to exist.