Export MIB nodes using their original names
MIB symbol names that contain hyphens cannot be used directly as
identifier names in the generated Python code. Therefore, such hyphens
are replaced with underscores when generating code for various MIB
nodes.

However, the generated export table also exports the symbols using the
"Pythonized" name as symbol name, rather than the original name. Before
the Jinja2 rewrite of git-a42d170, the original symbol names were then
made available by generating calls to setLabel() on MIB nodes. The label
generation got lost as part of that rewrite, with only some loose code
fragments remaining. That caused import errors for various MIBs.

Restoring the setLabel() calls would not be difficult. However, there is
no reason to export symbols using their Pythonized name at all: instead,
they can be exported by means of their original names from the MIB
files. The effect is the same, with less code and fewer exceptions to be
made. In addition, a next commit will enable exporting types. Types are
not covered by the existing setLabel() facility, and therefore would
always have required being exported by their real names.

Therefore, this commit exports the MIB nodes with their original names,
so that those nodes can be imported using those same names, as was
already generally expected. Therefore, this commit gets rid of the
aforementioned import errors, mainly by simplifying existing code.

As part of this change, the Jinja2 template is cleaned up, as it
performed many hyphen-to-underscore conversions that were entirely
redundant: with imported symbols as only exception, the symbol names are
always already converted to their Python versions before being passed to
the template. The template clean-up is necessary to avoid confusing the
reader as to which names are "Pythonized" or not already, and also
serves as the basis for another next commit (regarding Python keywords).

The existing tests are extended with new test cases to ensure that the
use of hyphens results in generation of proper Python code and that the
resulting MIB nodes' getLabel() method calls return the original names.

Please note that this commit makes a breaking change to the JSON output:
the "name" entries now contain the original, un-Pythonized version of
their corresponding entities. Note that in the previous JSON format, the
original node name from the MIB was not available at all.