Fix use of reserved Python keywords
This commit fixes issues related to the use of Python keywords in MIB
files, which cause problems in the generated Python code if left as is.
Support for translating such reserved symbols to non-reserved symbols
was already already in place to some degree, but inadequate, as there
were still issues for both defining and importing symbols.

For defined symbols, this change replaces the incomplete solution from
git-b091479b with the original fix of git-00177b06, which was still in
place for symtable.py but got dropped for intermediate.py as part of the
Jinja2 rewrite of git-a42d170.

For imported symbols, this change centralizes the symbol replacement and
ensures that the Jinja2 template generates an import table with local
symbol names that have been subject to appropriate replacement.

The commit adds various new tests for both cases.