Properly enquote, escape, preserve text values
Free-form text values, such as DESCRIPTION strings, may contain a few
characters that can be problematic when transformed to Python strings:

 1. Line terminator characters may be used in any textual strings, as
    per RFC 2578 Sec. 3.1.1. The Jinja2 template uses triple-quote
    strings to support such characters in some of those textual strings,
    but not in all of them. When line terminator characters are used
    within strings that are transformed to single-quote Python strings,
    the resulting Python code can not be compiled.

 2. Backslash characters, which are similarly allowed in such textual
    strings, are interpreted within Python strings. They should instead
    be converted to literal characters. Note that the final character of
    a string may be a backslash, so use of Python's "r" string prefix
    does not fully solve this problem.

This commit adds a Jinja2 filter that can be used to convert any MIB-
provided textual string to a Python string with proper quoting and
escaping of backslashes. That new filter is now used for all textual
strings. To be exact: it is now used for all strings that are ultimately
obtained via the QUOTED_STRING parser token, which is (or at least
should be) the only way through which free-form text values can make
their way from input MIBs into the code generator.

As direct effect of this change, the proper functionality of mibdump's
"--keep-texts-layout" switch is restored. Since the switch to Jinja2 of
commit git-a42d1705, triple-quoted lines had an extra line terminator
character added at the end, and were always getting word-wrapped. Both
aspects are dropped with this change, so that "--keep-texts-layout"
results in a compiled MIB that perfectly preserves the original text
(aside from the original line terminator encodings, i.e., LF vs CR/LF).

The existing tests have been changed to reflect the removal of the
extra line terminator character as added with git-a42d1705. A few extra
representative tests are added to ensure that the above two cases are
handled properly in general and that the implementation of the
"--keep-texts-layout" switch functions as expected.