Fix cyclic dependencies resulting in freezes
A MIB may attempt to import itself, and multiple MIBs may attempt to
import each other. The pysmi compiler deals well with such cases as long
as the MIB *file* name matches the MIB name *within* the file. However,
the latter name is used to track whether a specific MIB has already been
loaded (or failed to load), which means that if that name differs from
the MIB's file name, pysmi may end up trying to load the MIB with the
same file name more than once, which results in an endless loop of
trying to load the same MIB file(s) repeatedly. In such cases, the
"mibdump" tool will simply end up hanging forever.

This commit adds basic protection for such freezes by ensuring that
every MIB file is loaded at most once. Depending on the specific MIBs,
circular imports may result in either successful or failed compilation.
That is: this commit does not aim to ensure the best possible outcome in
such cases, as they are rare and not generally supported. However, this
commit does ensure that such cases do not block progress.