Add support for mistyped "SEQUENCE OF" type names
As per RFC 2578, conceptual-table object types (i.e., xyzTable) are
expected to have a "SEQUENCE OF" syntax whose type matches the syntax of
the table row (i.e., of xyzEntry). The pysmi symbol table building
procedure relies on that fact, and currently rejects any MIBs that do
not conform to that rule. However, there are several real-world MIBs
that have typos in their "SEQUENCE OF" type names.

This commit adds a leniency exception to the code in order to allow MIBs
with such typos to be compiled and loaded after all. The exception is
applied only when compiling the MIB would certainly fail otherwise, so
compilation of valid MIBs is guaranteed not to be affected by this
change. The exception requires keeping track of SEQUENCE types, but as
non-SEQUENCE types are already tracked as well, this is not considered a
substantial resource drain. A test case is added to verify the leniency.