Add support for SMIv1 index syntaxes
SMIv2 requires that an INDEX clause lists a series of object-type names.
Unlike SMIv2 however, SMIv1 (as per RFC 1212 Sec. 2) also supports that
such a series contains certain index syntaxes instead of (or in addition
to) object-type names. For example, SMIv1 allows "INDEX { INTEGER }".

While pysmi's parser supports the latter case, its code generator has
support in place in its symbol table builder only, with the intermediate
module merely having a few loose and non-functional code fragments with
the remainder being left as a to-do item.

This commit finishes the support for SMIv1 index syntaxes, by properly
generating "fake column" object-type instances for the index syntaxes,
as was likely the original intention as well.

A few duplicate class variables in the symbol table and intermediate
modules are moved to the base class, as successful Python code
generation relies on the fact that both subclasses use exactly the same
values for those. While there, a small set of unused and obsolete class
variables is removed altogether.

The test set is extended with a new test module that covers the new
SMIv1-only functionality.