Fix rewriting imports from RFC1158-MIB
As part of pysmi's code generation, various imports are rewritten from
SMIv1 to SMIv2, so as to simplify the way in which standard MIBs are
handled. However, an apparent copy-paste error prevents that imports
from RFC1158-MIB are being rewritten as intended. The result is that
MIBs importing from RFC1158-MIB could be compiled but not loaded.

This commit fixes the copy-paste error, and adds a few basic test cases
to ensure that each group of rewrites is working as intended.