Deal properly with IMPORTS being absent
Omitting the IMPORTS statement is valid, but pysmi does not deal with
this case as it should. When there is an IMPORTS statement, pysmi
extends the parsed imports table with necessary extra imports. When the
IMPORTS statement is omitted however, that extension is applied to a
temporary throw-away import table instead.

For example, when a MIB without an IMPORTS table uses the INTEGER type,
pysmi rewrites INTEGER to Integer32 as usual. However, since the MIB's
imports table is not extended in this case, Integer32 itself is never
imported from SNMPv2-SMI, resulting in a load-time error.

This commit ensures that there is always a parsed import table, even if
it is empty. The extensions are then applied to it in all cases, and the
above example works as intended. A test case is added as well.