Allow tables to be used as table columns
This fix is specifically for WIENER-CRATE-MIB, which uses a "SEQUENCE"
declaration that includes a table object symbol. As a result, the
corresponding object type is considered to be a table column, resulting
in generated Python code that cannot be loaded.

This commit adds an extra restriction for object types to be considered
table columns, namely that they must not have a "SEQUENCE OF" type
syntax. In other words, object types that are tables are no longer ever
considered to be table columns. That restriction allows WIENER-CRATE-MIB
to be loaded properly.

The commit also adds a test case, which consists of a simplified version
of the construction found in WIENER-CRATE-MIB.