Sort imported symbols in generated pysnmp code
For development, debugging, and regression testing purposes, it is
useful for code generation to be deterministic, in that when the same
(unchanged) MIB is compiled to Python code multiple times, the same code
is generated each time--at least, aside from the header comments.

The pysmi code generator currently already gets very close to that goal,
but falls short when it comes to the generated lists of symbols per
imported module. This commit makes generation of those lists
deterministic as well.