Make available named values of BITS used in types
When an OBJECT-TYPE definition uses BITS in its SYNTAX, the resulting
Python code will add the definition of the bits (each with name and
number) to the object as named values, as is intended and expected.
However, when an OBJECT-TYPE definition uses a SYNTAX that refers to a
simple type (for SMIv1) or a textual convention (for SMIv2) that uses a
BITS syntax, then the resulting Python code does *not* make available
the bits as named values in any way. Such "indirect" use of BITS via
types is however perfectly valid as well, and indeed used in practice.

Fortunately, this problem is limited to the Python code generation
template, which simply does not generate the named values for the typed
(indirect) case. The necessary information is available internally.

This commit extends the template to generate the named values for
BITS-type simple types and textual conventions. That also makes the
generation consistent with similar constructions already present in the
template (e.g., for generation of named values for enumerated INTEGERs).
The change affects only MIBs whose generated Python code was missing the
named values for BITS.

New unit test cases are added for the cases that previously failed.