Make more text load only when requested
The pysnmp 'loadTexts' setting controls whether text-based metadata
should be assigned to objects at load time of compiled MIBs. Currently,
in a few cases, text-based metadata is being assigned to objects even
when 'loadTexts' is disabled. Before the switch to Jinja2, those cases
were however properly covered by a 'loadTexts' check, so that appears to
have been a regression. This commit changes the assignments such that
they are once again behind a 'loadTexts' check.

As part of this commit, the test set is expanded to test that the
resulting 'loadTexts' behavior is as intended, both for more cases in
which 'loadTexts' is enabled (completing the extensions introduced in
git-1e4172e), and many new cases in which 'loadTexts' is disabled.

Please note that pysmi's own 'genTexts' setting adds an extra dimension
with respect to (not) loading texts. Neither the pysmi change nor the
test extensions in this commit address 'genTexts' in any way; that is
left to future work.

While working on the test set, also rename the "traptype" test from
"smiv2" to "smiv1", as it uses exclusively SMIv1 constructions.