Fix DEFVAL for inherited enumerations
If an object type uses an enumerated-INTEGER syntax, and also specifies
a DEFVAL (with a value from the enumeration), then that default value is
also properly assigned to the object type's class in the generated
Python code. That is the expected behavior.

However, if an object type uses an "indirect" syntax instead, using a
type (e.g., a TEXTUAL-CONVENTION) that in turn is based on an enumerated
INTEGER, then the same case results in the default value being omitted
from the generated Python code. That is a bug.

The bug is a result of the Jinja2 template being given the enumeration
*name* of the default value, and having to do the name-to-value lookup
itself. That lookup assumed that the object type was also the entity
that defined the enumeration, which (as per the second case above) is
not always true.

This commit changes the information handed to the Jinja2 template to the
enumeration *value* of the default value (i.e., a number, as string), so
that the template no longer needs to do a lookup at all. The
intermediate code is in a better position to do the proper name-to-value
lookup, without making assumptions on where the enumeration is defined.

The overall result is that many more object types from many MIBs get
their intended default values assigned to them.

The test set is extended to cover the most typical occurrences of these
aspects, including the case that the type that defines the enumeration
comes from another MIB altogether (which itself already worked as
intended, but would be easy to break in the future). Also included is a
test case for a related minor bug that was automatically resolved as
part of this change, even though it is based on an illegal construction
(namely, of range-subtyping an enumerated INTEGER).

Please note that this commit changes the pysmi JSON output format.