Always override intermediate syntax class names
When an OBJECT-TYPE declaration subtypes a syntax, uses a BITS syntax,
or uses a DEFVAL clause, the MIB definitions template generates an
intermediate syntax type class named _Xyz_Type for object Xyz. For the
first two of those three cases, the template then overrides the class's
__name__ field to reflect the name of the original syntax. However, for
cases where only the third (DEFVAL) case applies, it does not. The
result is that users that obtain getSyntax().__class__.__name__ will
often, but not always, find a proper syntax type name; in the DEFVAL
case, they will get the internal "_Xyz_Type" name instead.

This change overrides the intermediate syntax type class's __name__
field in all of the three cases, so that __name__ always yields a name
that is indeed a name of a valid syntax type class higher up in the
class hierarchy of the corresponding object.

A corresponding test set is added to cover various local (intra-MIB) and
import (inter-MIB) cases.