Fix registration of ModuleCompliance objects
ModuleCompliance instances were erroneously not assigned any objects,
meaning that getObjects() would always return an empty list. The reason
for that problem is the use of a wrong field name in the Jinja2
template for pysnmp code generation.

This commit fixes that bug, and adds unit tests for the classes that
support objects but did not yet have unit tests for those.