Improve on the use of symbol "Pythonization"
Given that pysmi uses many ASN.1 symbol names directly as Python symbol
names in generated code, it must convert some of those symbols to make
for usable Python names. For example, hyphens are converted to
underscores, and reserved Python keywords (such as "if" and "global")
are prefixed with "_pysmi_" to avoid collisions. Such "Pythonization" of
ASN.1 symbol names is done internally by trans_opers().

For by far most symbols, trans_opers() need not apply changes. That has
resulted in the situation that pysmi has many bugs where trans_opers()
is erroneously (not) invoked. One result of that is that users of pysnmp
are exposed to the Pythonized names rather than the original ASN.1 names
in various cases, making those unusable for subsequent pysnmp API calls.

This commit rectifies the use of trans_opers() for the following cases:

  o Object identifier type DEFVALs. Use of OID names that required
    "Pythonization" would result in failed compilation.
  o Object lists of NotificationType, ObjectGroup, NotificationGroup,
    TrapType, and ModuleCompliance classes. Those would previously
    contain Pythonized names, as returned by their getObjects() APIs.
  o Table columns. Columns with names that required "Pythonization"
    would not be exposed as MibTableColumn subclasses.
  o External table indices. Indices that required "Pythonization" would
    be listed as coming from the local MIB rather than the MIB from
    which they are imported, as returned by getIndexNames().
  o Table index augmentation registrations. Those would previously
    contain Pythonized names. Note that there is (still) no pysnmp API
    to obtain such registrations.

Additional clean-up removes a case where trans_opers() was called twice
on the same name, and removes the unused tracking of columns from the
intermediate code generation stage. A few more comments are added to
clarify which internal names are Pythonized and which are not.

The test set is extended to cover all of the above cases. In the future,
the "Pythonization" should be subjected to a more thorough overhaul, as
the current approach is hard to follow and maintain. Hopefully, the new
tests will make such an overhaul a bit easier to get right.