Register fixed lengths for octet strings
For the purpose of generating and parsing formatted OIDs, pysnmp needs
to know whether an OCTET STRING typed object has a fixed length, and if
so, which length. Before the switch to Jinja2 templates, pysmi assigned
such fixed lengths as "fixedLength" attributes to types. With the
switch, that functionality got lost, thereby also breaking pysnmp's
proper handling of formatted OIDs.

This commit restores the assignment of "fixedLength" attributes. Note
that as of writing, this change is necessary but not sufficient to fix
the corresponding funtionality in pysnmp.

The test set is extended with extra tests for this functionality, and
some previously missing general tests on constraints. As part of that, a
bug is fixed in an existing test wherein naming conflicts caused
various bulk-generated test methods to overwrite each other.