Improve processing of DEFVAL values
This commit aims to define clearly which combinations of DEFVAL value
types and underlying syntax types are supported, and to implement
support for those cases. Various bugs in the previous code are resolved
as well. The overall result is that the generated Python code provides
proper default values for object types in more cases than before, and
that it no longer assigns values that are not used or usable by pyasn1.

Concretely, the following DEFVAL-related bugfixes are applied:

  - Empty quoted strings are no longer discarded;
  - Decimal values are no longer assigned to string-type and OID-type
    objects, which used to lead to corrupted default values;
  - Integer values with hex/binary notation are no longer assigned
    incorrectly, previously resulting in them being discarded;
  - For strings, binary notation no longer discards leading zero octets;
  - Leading zeroes are added to hex/binary notation values with invalid
    lengths, which often used to result in unintended default values;
  - Numeric values for enumerated integers that fall outside of the
    permitted enumeration values are now discarded;
  - Code generation no longer produces extra unnecessary _Xyz_Type type
    classes without contents.

In addition, the code is changed so that it no longer fails compiling
MIBs for any case in which a DEFVAL value can be parsed but not
processed. The result is that fewer MIBs fail compilation altogether.

As part of this commit, the functions that process default values are
restructured, primarily to filter by the syntax base type first and only
then by the input type, rather than the other way around. That change
opens up the possibility of applying additional checks to the default
values in a central place, as will be done in a follow-up commit.

The code is cleaned up somewhat as well. The intermediate code no longer
uses the same gen_def_val() method for two different purposes. The
template code now has fewer cases handling default values. Unused DEFVAL
code is removed from the symtable code. A few Python2 remnants are
cleaned up as well.

The test set is extended with many more DEFVAL test cases. As part of
that, all the DEFVAL-specific tests are moved to a new module, as their
original location was getting too large.

This commit slightly alters the JSON output format.