Apply constraints checks to DEFVAL values
For DEFVAL clauses associated with object type syntaxes that are:
  1. derived from octet strings or (non-enumerated) integers, and,
  2. sub-typed with value range or size constraints,
..the DEFVAL values are now checked against those constraints, and
discarded if they do not adhere to the constraints.

As a result, any DEFVAL values that would otherwise cause a MIB loading
failure due to pyasn1 value constraint errors, are now kept out of the
generated Python code altogether. That allows many more slightly buggy
MIBs to be loaded and used than before, albeit without the offending
DEFVAL values of course.

The test set is extended to cover the new functionality.